Engine API: Split const evaluation into a separate file
This is a preparation for the future removal of Runtime.swift.
